backend/requirements.txt
motor==3.3.2
python-dotenv==1.0.0
pydantic[email]==2.5.0
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic-settings==2.1.0
python-jose[cryptography]==3.3.0
passlib==1.7.4
python-multipart==0.0.6
email-validator==2.1.0
bcrypt==4.0.1
aiohttp==3.9.1